Add webApp module to follow the project convention
