+++ compose-imageviewer/shared/src/commonMain/kotlin/example/imageviewer/ImageViewer.common.kt
@@ -34,7 +34,7 @@ enum class ExternalImageViewerEvent {
 
 @OptIn(ExperimentalAnimationApi::class)
 @Composable
-internal fun ImageViewerCommon(
+fun ImageViewerCommon(
     dependencies: Dependencies,
     externalEvents: Flow<ExternalImageViewerEvent> = emptyFlow()
 ) {
