+++ compose-imageviewer/webApp/src/jsMain/kotlin/Main.kt
@@ -1,0 +1,19 @@
+import androidx.compose.runtime.*
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.CanvasBasedWindow
+import example.imageviewer.*
+import example.imageviewer.model.*
+import io.ktor.client.*
+import io.ktor.client.engine.js.*
+import io.ktor.client.request.*
+import io.ktor.client.statement.*
+import org.jetbrains.skiko.wasm.onWasmReady
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    onWasmReady {
+        CanvasBasedWindow("ImageViewer") {
+            ImageViewerWeb()
+        }
+    }
+}
