+++ compose-imageviewer/webApp/src/jsMain/resources/index.html
@@ -6,11 +6,7 @@
     <script src="skiko.js"> </script>
 </head>
 <body>
-<center>
-    <div style="background-color: darkslategrey;">
-        <canvas id="ComposeTarget" width="800" height="1000"></canvas>
-    </div>
-</center>
+<canvas id="ComposeTarget"></canvas>
 <script src="imageviewer.js"> </script>
 </body>
 </html>
