+++ compose-imageviewer/webApp/src/jsWasmMain/kotlin/Common.kt
@@ -2,12 +2,9 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.Surface
 import androidx.compose.runtime.*
 import androidx.compose.runtime.snapshots.SnapshotStateList
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.window.Window
 import example.imageviewer.*
-import example.imageviewer.ImageViewerCommon
 import example.imageviewer.core.BitmapFilter
 import example.imageviewer.core.FilterType
 import example.imageviewer.model.*
@@ -20,25 +17,8 @@ import example.imageviewer.view.Toast
 import example.imageviewer.view.ToastState
 import kotlinx.coroutines.CoroutineScope
 import org.jetbrains.compose.resources.ExperimentalResourceApi
-import org.jetbrains.compose.resources.Resource
-import org.khronos.webgl.ArrayBuffer
-import org.khronos.webgl.Int8Array
-import org.w3c.xhr.XMLHttpRequest
-import org.w3c.xhr.XMLHttpRequestResponseType
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
-import kotlin.coroutines.suspendCoroutine
-import kotlin.wasm.unsafe.*
-import kotlin.wasm.unsafe.withScopedMemoryAllocator
-import kotlin.wasm.unsafe.UnsafeWasmMemoryApi
-import androidx.compose.ui.window.CanvasBasedWindow
 
-@OptIn(ExperimentalComposeUiApi::class)
-fun main() {
-    CanvasBasedWindow("ImageViewer") {
-        ImageViewerWeb()
-    }
-}
+expect fun createWrappedHttpClient(): WrappedHttpClient
 
 @Composable
 internal fun ImageViewerWeb() {
@@ -84,12 +64,8 @@ fun getDependencies(ioScope: CoroutineScope, toastState: MutableState<ToastState
         override val back = "Back"
     }
 
-    override val httpClient: WrappedHttpClient = object : WrappedHttpClient {
 
-        override suspend fun getAsBytes(urlString: String): ByteArray {
-            return urlResource(urlString).readBytes()
-        }
-    }
+    override val httpClient: WrappedHttpClient = createWrappedHttpClient()
 
     private fun ContentRepository<ImageBitmap>.cacheByUrlAdapter(): ContentRepository<ImageBitmap> {
         val original = this
@@ -112,77 +88,5 @@ fun getDependencies(ioScope: CoroutineScope, toastState: MutableState<ToastState
         override fun showPopUpMessage(text: String) {
             toastState.value = ToastState.Shown(text)
         }
-    }
-}
-
-@ExperimentalResourceApi
-private fun urlResource(path: String): Resource = JSResourceImpl(path)
-
-@OptIn(ExperimentalResourceApi::class)
-private abstract class AbstractResourceImpl(val path: String) : Resource {
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        return if (other is AbstractResourceImpl) {
-            path == other.path
-        } else {
-            false
-        }
-    }
-
-    override fun hashCode(): Int {
-        return path.hashCode()
-    }
-}
-
-@ExperimentalResourceApi
-private class JSResourceImpl(path: String) : AbstractResourceImpl(path) {
-    override suspend fun readBytes(): ByteArray {
-        return suspendCoroutine { continuation ->
-            val req = XMLHttpRequest()
-            req.open("GET", path, true)
-            req.responseType = "arraybuffer".asDynamic().unsafeCast<XMLHttpRequestResponseType>()
-
-            req.onload = { _ ->
-                val arrayBuffer = req.response
-                if (arrayBuffer is ArrayBuffer) {
-                    val size = arrayBuffer.byteLength
-                    continuation.resume(arrayBuffer.toByteArray())
-                } else {
-                    continuation.resumeWithException(MissingResourceException(path))
-                }
-                null
-            }
-            req.send(null)
-        }
-    }
-}
-
-private class MissingResourceException constructor(path: String) :
-    Exception("Missing resource with path: $path")
-
-
-private fun ArrayBuffer.toByteArray(): ByteArray {
-    val source = Int8Array(this, 0, byteLength)
-    return jsInt8ArrayToKotlinByteArray(source)
-}
-
-@JsFun(
-    """ (src, size, dstAddr) => {
-        const mem8 = new Int8Array(wasmExports.memory.buffer, dstAddr, size);
-        mem8.set(src);
-    }
-"""
-)
-internal external fun jsExportInt8ArrayToWasm(src: Int8Array, size: Int, dstAddr: Int)
-
-internal fun jsInt8ArrayToKotlinByteArray(x: Int8Array): ByteArray {
-    val size = x.length
-
-    @OptIn(UnsafeWasmMemoryApi::class)
-    return withScopedMemoryAllocator { allocator ->
-        val memBuffer = allocator.allocate(size)
-        val dstAddress = memBuffer.address.toInt()
-        jsExportInt8ArrayToWasm(x, size, dstAddress)
-        ByteArray(size) { i -> (memBuffer + i).loadByte() }
     }
 }
