+++ compose-imageviewer/webApp/src/wasmMain/resources/index.html
