+++ compose-imageviewer/webApp/src/wasmMain/resources/load.mjs
